Working TCP cancellation, command rejection, trajectory sending. Ready for integration with online BCI scenario

+++ nav_goals/include/nav_goals/commandreader.h
@@ -36,7 +36,7 @@ private:
     int current_stim;
     int left_stim_counter;
     int right_stim_counter;
-    int counter_weight;
+    double counter_weight;
     int direction;
     int orientation;
